Rename endpoints
