+++ design/media_types.go
@@ -5,24 +5,6 @@ import (
 	a "github.com/goadesign/goa/design/apidsl"
 )
 
-// ALMStatus defines the status of the current running ALM instance
-var ALMStatus = a.MediaType("application/vnd.status+json", func() {
-	a.Description("The status of the current running instance")
-	a.Attributes(func() {
-		a.Attribute("commit", d.String, "Commit SHA this build is based on")
-		a.Attribute("buildTime", d.String, "The time when built")
-		a.Attribute("startTime", d.String, "The time when started")
-		a.Attribute("error", d.String, "The error if any")
-		a.Required("commit", "buildTime", "startTime")
-	})
-	a.View("default", func() {
-		a.Attribute("commit")
-		a.Attribute("buildTime")
-		a.Attribute("startTime")
-		a.Attribute("error")
-	})
-})
-
 var pagingLinks = a.Type("pagingLinks", func() {
 	a.Attribute("prev", d.String)
 	a.Attribute("next", d.String)
