+++ design/status.go
@@ -1,0 +1,46 @@
+package design
+
+import (
+	d "github.com/goadesign/goa/design"
+	a "github.com/goadesign/goa/design/apidsl"
+)
+
+// AuthStatus defines the status of the current running Auth instance
+var AuthStatus = a.MediaType("application/vnd.status+json", func() {
+	a.Description("The status of the current running instance")
+	a.Attributes(func() {
+		a.Attribute("commit", d.String, "Commit SHA this build is based on")
+		a.Attribute("buildTime", d.String, "The time when built")
+		a.Attribute("startTime", d.String, "The time when started")
+		a.Attribute("error", d.String, "The error if any")
+		a.Required("commit", "buildTime", "startTime")
+	})
+	a.View("default", func() {
+		a.Attribute("commit")
+		a.Attribute("buildTime")
+		a.Attribute("startTime")
+		a.Attribute("error")
+	})
+})
+
+var _ = a.Resource("status", func() {
+
+	a.DefaultMedia(AuthStatus)
+	a.BasePath("/status")
+
+	a.Action("show", func() {
+		a.Routing(
+			a.GET(""),
+		)
+		a.Description("Show the status of the current running instance")
+		a.Response(d.OK)
+		a.Response(d.ServiceUnavailable, AuthStatus)
+	})
+})
+
+var nameValidationFunction = func() {
+	a.MaxLength(62) // maximum name length is 62 characters
+	a.MinLength(1)  // minimum name length is 1 characters
+	a.Pattern("^[^_|-].*")
+	a.Example("name for the object")
+}
